+++ Src/CommonStates.cpp
@@ -391,7 +391,7 @@ public:
     static const D3D12_SAMPLER_DESC SamplerDescs[static_cast<int>(SamplerIndex::Count)];
 
     Impl(_In_ ID3D12Device* device)
-        : mDescriptors(device, D3D12_DESCRIPTOR_HEAP_TYPE_SAMPLER, D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE, (int)SamplerIndex::Count)
+        : mDescriptors(device, D3D12_DESCRIPTOR_HEAP_TYPE_SAMPLER, D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE, static_cast<size_t>(SamplerIndex::Count))
     {
         SetDebugObjectName(mDescriptors.Heap(), L"CommonStates");
 
@@ -403,7 +403,7 @@ public:
 
     D3D12_GPU_DESCRIPTOR_HANDLE Get(SamplerIndex i) const
     {
-        return mDescriptors.GetGpuHandle((int)i);
+        return mDescriptors.GetGpuHandle(static_cast<size_t>(i));
     }
 
     ID3D12DescriptorHeap* Heap() const
